Create: Add function & variable to script.js
I added a function getComputerChoice that takes the value of randomInt
and converts it to one of three strings; "Rock", "Paper", or "Scissors".
I did this using if/if else conditional syntax within the function call.
The value stored in randomInt can only be a 1, 2, or 3, so I included a
contitional test for each number, thereby transforming a random number
generator into one of three random strings.

I then added a variable computerChoice to store the return output of the
getComputerChoice function.