Feat: Add function to script.js
I took the code for the function playRound that I worked out in the
pseudocoding stage and copied it in to script.js. The function accepts
two parameters, playerSelection and computerSelection, and includes a
series of if/else if statements.

It first checks for strict equality
between to two, to determine if a draw has taken place. Subsequent
else if statements use the && logical operator to determine
relationships between each parameter and whether the game is a win or
a loss.