Feat: Add const & console.log function to .js file
In order to test the playRound function I added the variable
playerSelection with the value "Rock" and a console.log function to call
the outcome of the game to the console.

The const value "Rock" is intentially lowercase to test the toUpperCase
functionality for the purpose of comparisonsons.

This was successful as the game functions exactly as intended. I double
checked this by changing the playerSelection value to "Paper" and
"Scissors" respectively.