Feat: Add function to script.js
I declared the function finalResult, which uses if/else if syntax to
display one of three messages to announce the result of the game, along
with a final tally of how many games were won by each side.

The function compares the values stored in playerWin and computerWin
using greater than, less than, or equals to operators.

The incrementation of these variables in the keepScore function reflect
the win/loss/draw results. I called their values within the text
strings that contain the final result message that is stored in this
finalResult function, ready to be called to the console at the end of
the game.