Fix: Fix output randomisation issue
Well that didn't take me long. I decided to try declaring the variables
randomInt and computerSelection along with their associated function
calls in all four instances nested in the game function declaration, as
opposed to simply recalling the getRandomInt and getComputerChoice
functions on those same lines as I had previously. Tested it, and the
fix was successful!

This means that the project outline has come to fruition. I will be
revisiting this exercise/repo in the future to implement G.U.I
functionality.