Change: Clean up code structure & edit return text
I decided to clean up the code for this program by consolidating
all function and variable declarations and function calls into a single
function: 'game'

This program previously declared/initialised variables and functions in
global scope, and then called the game function to play four rounds.
This is because I needed to initialise functions and variables in global
scope, which would inadvertedly invoke the core functions and play a
single round before calling 'game' to play the remaining rounds.

This was a result of my initial troubleshooting workflow, so I decided
to come back to this project with fresh eyes and clean up the code.

I also slightly edited the return strings in the finalResult function in
order to better differentiate the final result message from the single
round result messages.